# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StockIQ — WMS API",
    "version": "1.0.0",
    "description": "API REST do StockIQ (WMS / ERP Lite).\n\n**Autenticação:** faça `POST /api/auth/login`, copie o `token` retornado e clique em **Authorize** (esquema `bearerAuth`).\n\n**Perfis:** `admin` > `supervisor` > `operator` > `viewer`. Cada operação lista os perfis permitidos.\n\n**Erros:** respostas de erro seguem `{ \"message\": string }`; validações retornam 422 com `errors[]`."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Descrição do erro"
          }
        },
        "required": [
          "message"
        ]
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Validation error"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "Invalid email"
                }
              }
            }
          }
        },
        "required": [
          "message"
        ]
      },
...
      "Movement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmq1x2y3z0001abcd"
          },
          "type": {
            "type": "string",
            "enum": [
              "entry",
              "exit",
              "transfer",
              "loss",
              "adjustment",
              "inventory"
            ],
            "example": "entry"
          },
          "quantity": {
            "type": "integer",
            "example": 50
          },
          "unitCost": {
            "type": "number",
            "example": 22.5
          },
          "totalValue": {
            "type": "number",
            "example": 1125
          },
          "invoiceNumber": {
            "type": "string",
            "nullable": true
          },
          "lotNumber": {
            "type": "string",
            "nullable": true
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "exitReason": {
            "type": "string",
            "enum": [
              "sale",
              "transfer",
              "loss",
              "break",
              "internal"
            ],
            "example": "sale",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "productId": {
            "type": "string",
            "example": "cmq1x2y3z0001abcd"
          },
          "supplierId": {
            "type": "string",
            "nullable": true
          },
          "customerId": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "fromAddressId": {
            "type": "string",
            "nullable": true
          },
          "toAddressId": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "example": "cmq1x2y3z0001abcd"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "product": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "name": {
                "type": "string"
              },
              "internalCode": {
                "type": "string"
              }
            }
          },
          "supplier": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "name": {
                "type": "string",
                "example": "Nome"
              }
            },
            "nullable": true
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "name": {
                "type": "string",
                "example": "Nome"
              }
            }
          },
          "fromAddress": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "code": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "toAddress": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "code": {
                "type": "string"
              }
            },
            "nullable": true
          }
        }
      },
...
      "Lot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmq1x2y3z0001abcd"
          },
          "lotNumber": {
            "type": "string",
            "example": "LOT-QM0091-001"
          },
          "quantity": {
            "type": "integer",
            "example": 200
          },
          "manufacturingDate": {
            "type": "string",
            "format": "date-time"
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "type": "string",
            "example": "C-03-02-A-01"
          },
          "status": {
            "type": "string",
            "enum": [
              "valid",
              "expiring",
              "expired",
              "quarantine"
            ],
            "example": "valid",
            "description": "Derived from expirationDate (≤30 days = expiring) unless the lot is in quarantine."
          },
          "productId": {
            "type": "string",
            "example": "cmq1x2y3z0001abcd"
          },
          "supplierId": {
            "type": "string",
            "example": "cmq1x2y3z0001abcd"
          },
          "product": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "name": {
                "type": "string"
              },
              "internalCode": {
                "type": "string"
              }
            }
          },
          "supplier": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cmq1x2y3z0001abcd"
              },
              "name": {
                "type": "string",
                "example": "Nome"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaginatedMovement": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Movement"
            }
          },
          "total": {
            "type": "integer",
            "example": 42
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "totalPages": {
            "type": "integer",
            "example": 3
          }
        }
      },
      "PaginatedLot": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Lot"
            }
          },
          "total": {
            "type": "integer",
            "example": 42
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "totalPages": {
            "type": "integer",
            "example": 3
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Token ausente, inválido ou expirado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Perfil sem permissão para esta operação",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Registro não encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflito (valor único já existente ou registro vinculado)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Erro de validação do corpo/parâmetros",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/movements": {
      "get": {
        "summary": "List movements",
        "tags": [
          "Movements"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "example": "exit,loss"
            },
            "description": "One movement type or a comma-separated list (entry, exit, transfer, loss, adjustment, inventory)"
          },
          {
            "in": "query",
            "name": "productId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Matches product name/code, operator, invoice, lot or customer"
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of movements",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMovement"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "post": {
        "summary": "Register a stock movement",
        "description": "Available to admins, supervisors, and operators. Viewers are read-only.\n\n**Perfis permitidos:** admin, supervisor, operator.",
        "tags": [
          "Movements"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "productId",
                  "quantity"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "entry",
                      "exit",
                      "transfer",
                      "loss",
                      "adjustment",
                      "inventory"
                    ]
                  },
                  "productId": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "unitCost": {
                    "type": "number"
                  },
                  "exitReason": {
                    "type": "string",
                    "enum": [
                      "sale",
                      "transfer",
                      "loss",
                      "break",
                      "internal"
                    ]
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Movement registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movement"
                }
              }
            }
          },
          "400": {
            "description": "Insufficient stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/movements/{id}": {
      "get": {
        "summary": "Get movement by ID",
        "tags": [
          "Movements"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movement found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movement"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/lots": {
      "get": {
        "summary": "List lots",
        "tags": [
          "Lots"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "productId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "valid",
                "expiring",
                "expired",
                "quarantine"
              ]
            }
          },
          {
            "in": "query",
            "name": "expiringSoonDays",
            "schema": {
              "type": "integer",
              "example": 30
            },
            "description": "Return lots expiring within this many days"
          }
        ],
        "responses": {
          "200": {
            "description": "List of lots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedLot"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "post": {
        "summary": "Register a new lot",
        "tags": [
          "Lots"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "lotNumber",
                  "productId",
                  "supplierId",
                  "quantity",
                  "manufacturingDate",
                  "expirationDate",
                  "address"
                ],
                "properties": {
                  "lotNumber": {
                    "type": "string"
                  },
                  "productId": {
                    "type": "string"
                  },
                  "supplierId": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer"
                  },
                  "manufacturingDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "expirationDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "address": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "valid",
                      "expiring",
                      "expired",
                      "quarantine"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lot created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lot"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "description": "**Perfis permitidos:** admin, supervisor, operator."
      }
    },
    "/api/lots/alerts": {
      "get": {
        "summary": "Get lots expiring soon (default 30 days)",
        "tags": [
          "Lots"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lots expiring soon",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lot"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/lots/{id}": {
      "get": {
        "summary": "Get lot by ID",
        "tags": [
          "Lots"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lot found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lot"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Update lot",
        "tags": [
          "Lots"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "valid",
                      "expiring",
                      "expired",
                      "quarantine"
                    ]
                  },
                  "quantity": {
                    "type": "integer"
                  }
                }
              }
            }
          },
          "description": "Todos os campos são opcionais na atualização parcial."
        },
        "responses": {
          "200": {
            "description": "Lot updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lot"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "description": "**Perfis permitidos:** admin, supervisor."
      },
      "delete": {
        "summary": "Delete a lot",
        "tags": [
          "Lots"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Lot deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "description": "**Perfis permitidos:** admin."
      }
    },
...
  }
}